Move the custom-inflector-rules.js file
The model folder should only include actual Models.
