+++ app/config/custom-inflector-rules.js
@@ -62,5 +62,3 @@ inflector.irregular('blockchain-status', 'blockchain-statuses');
 inflector.irregular('concept', 'concepts');
 inflector.irregular('concept-scheme', 'concept-schemes');
 inflector.uncountable('sync');
-// Meet Ember Inspector's expectation of an export
-export default {};
